Restructuring models.  Adding troop app to manage troop information.
