+++ mbu/admin.py
@@ -1,12 +1,9 @@
 from django.contrib import admin
-from mbu.models import Troop, TroopContact, Address, Counselor, Venture, MeritBadgeUniversity, WaitingList
+from mbu.models import Address, Counselor, Venture, MeritBadgeUniversity
 
-admin.site.register(Troop)
-admin.site.register(TroopContact)
 admin.site.register(Address)
 admin.site.register(Counselor)
 admin.site.register(Venture)
 admin.site.register(MeritBadgeUniversity)
-admin.site.register(WaitingList)
 
 # Register your models here.
