+++ registration/admin.py
@@ -1,4 +1,1 @@
 from django.contrib import admin
-from mbu.models import Scout
-
-admin.site.register(Scout)
