+++ scout/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
+from scout.models import Scout
 
 # Register your models here.
+
+admin.site.register(Scout)
