+++ scout/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
+from troop.models import Troop
 
 # Create your models here.
 
@@ -7,4 +8,4 @@ class Scout(models.Model):
     user = models.OneToOneField(User)
     dob = models.DateTimeField()
     rank = models.CharField(max_length=15)
-    troop = models.CharField(max_length=15)
+    troop = models.ForeignKey(Troop)
