+++ troop/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from troop.models import Troop, Council, TroopContact
+
+# Register your models here.
+admin.site.register(Troop)
+admin.site.register(Council)
+admin.site.register(TroopContact)
